+++ cooked-validators/src/Cooked/MockChain.hs
@@ -11,6 +11,5 @@ where
 import Cooked.MockChain.Monad as X
 import Cooked.MockChain.Monad.Direct as X
 import Cooked.MockChain.Monad.Staged as X
-import Cooked.MockChain.RawUPLC as X
 import Cooked.MockChain.Testing as X
 import Cooked.MockChain.UtxoState as X
